--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-layout-validator/.classpath
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<classpath>
-	<classpathentry excluding="**/.svn/**|.svn/" kind="src" path="src" />
-	<classpathentry kind="src" path="/portal-master" />
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER" />
-	<classpathentry excluding="**/.svn/**|.svn/" kind="src" path="test/unit" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/junit.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/mockito.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/powermock-api-mockito.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/powermock-api-support.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/powermock-core.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/powermock-module-junit4.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/powermock-module-junit4-common.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/spring-test.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/activation.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/annotations.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/jsp-api.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/mail.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/development/servlet-api.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/global/portlet.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/portal/bnd.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/portal/commons-io.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/portal/commons-lang.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/portal/commons-logging.jar" />
-	<classpathentry kind="lib" path="/portal-master/lib/portal/log4j.jar" />
-	<classpathentry kind="lib" path="/portal-master/portal-service/portal-service.jar" />
-	<classpathentry kind="lib" path="/portal-master/util-bridges/util-bridges.jar" />
-	<classpathentry kind="lib" path="/portal-master/util-java/util-java.jar" />
-	<classpathentry kind="lib" path="/portal-master/util-taglib/util-taglib.jar" />
-	<classpathentry kind="lib" path="/portal-master/.ivy/cache/org.hamcrest/hamcrest-core/jars/hamcrest-core-1.3.jar" />
-	<classpathentry kind="lib" path="/portal-master/.ivy/cache/junit/junit/jars/junit-4.11.jar" />
-	<classpathentry kind="output" path="bin" />
-</classpath>
