--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-layout-validator/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Dynamic Data Mapping Layout Validator
-Bundle-SymbolicName: com.liferay.dynamic.data.mapping.validator
-Bundle-Version: 1.0.0
-Include-Resource:\
-	classes,\
-	META-INF/resources=src/META-INF/resources
