--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-layout-validator/ivy.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-
-<ivy-module
-	version="2.0"
-	xmlns:m2="http://ant.apache.org/ivy/maven"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:noNamespaceSchemaLocation="http://ant.apache.org/ivy/schemas/ivy.xsd"
->
-	<info module="${plugin.name}" organisation="com.liferay">
-		<extends extendType="configurations,description,info" location="${sdk.dir}/ivy.xml" module="com.liferay.sdk" organisation="com.liferay" revision="latest.integration" />
-	</info>
-
-	<publications>
-		<artifact type="jar" />
-		<artifact type="pom" />
-		<artifact m2:classifier="sources" />
-	</publications>
-
-	<dependencies defaultconf="default">
-		<dependency conf="test->default" name="junit" org="junit" rev="4.11" />
-	</dependencies>
-</ivy-module>
