+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-validator/test/unit/com/liferay/dynamic/data/mapping/validator/DDMFormLayoutValidatorTest.java
@@ -21,7 +21,6 @@ import com.liferay.portlet.dynamicdatamapping.model.DDMFormLayoutColumn;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormLayoutPage;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormLayoutRow;
 import com.liferay.portlet.dynamicdatamapping.model.LocalizedValue;
-import com.liferay.portlet.dynamicdatamapping.validator.DDMFormLayoutValidator;
 
 import java.util.Locale;
